#2605: BsV: Custom biostructure data provider, bio
bio: Add biostructureIdColumnName, biostructureDataProvider
