+++ libraries/bio/package.json
@@ -8,7 +8,7 @@
     "access": "public"
   },
   "friendlyName": "Datagrok bio library",
-  "version": "5.39.14",
+  "version": "5.39.15",
   "description": "",
   "dependencies": {
     "@datagrok-libraries/gridext": "^1.3.68",
